+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Added
 - Added the option to exclude DHW from the peak heating load (issue #252).
 
+## Changed
+- Change skopt to Optuna in active_passive_cooling (issue #258).
+
 ## [2.2.2] - 2024-05-16
 
 ## Added